Calculate Blas_Job::cost in constructor
